requirements.txt
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
opencv-python>=4.9.0
Pillow>=10.0.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
streamlit>=1.30.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
requests>=2.31.0
python-docx>=1.1.0
kaleido>=1.0.0
paddlepaddle>=3.0.0
paddleocr>=2.9.0
